(DCMAW-9609): fix code smells and security hotspots (#72)
* feat(DCMAW): try to fix security hotstop in Dockerfile

* feat(DCMAW): fix code smell in middleware

* feat(DCMAW): fix code smells

* feat(DCMAW): fix code smells

* feat(DCMAW): add RetentionInDays to WalletFEAPIGWAccessLogsGroup

* build(DCMAW-9609): allow only SSL requests in WalletFEAccessLogsBucket

* build(DCMAW-9609): try to fix dockerfile error

+++ src/services/kmsService.ts
@@ -69,7 +69,7 @@ export class KmsService {
       throw new Error("No public key returned");
     }
 
-    return this.parsePublicKey(response.PublicKey as Uint8Array);
+    return this.parsePublicKey(response.PublicKey);
   }
 
   private parsePublicKey(publicKeyRaw: Uint8Array) {
